+++ apps/router-e2e/app.config.js
@@ -34,6 +34,7 @@ module.exports = {
     backgroundColor: '#ffffff',
   },
   experiments: {
+    autolinkingModuleResolution: true,
     baseUrl: process.env.EXPO_E2E_BASE_PATH || undefined,
     tsconfigPaths: process.env.EXPO_USE_PATH_ALIASES,
     typedRoutes: true,